semana09/dia3/backend/api_django/requirements.txt
django==3.2
djangorestframework
mysqlclient
